# Reduced NCNN ARM64 thread usage