## 🛠️ PR Summary

<sub>Made with ❤️ by [Ultralytics Actions](https://github.com/ultralytics/actions)<sub>

### 🌟 Summary
Optimizes multi-threading configuration for faster edge inference on ARM64 devices.

### 📊 Key Changes
- Adjusted threading settings in the edge inference code to use only the "big" CPU cores on ARM64 architectures.
- Added `NUM_THREADS` import to the autobackend script.

### 🎯 Purpose & Impact
- **Improved Performance**: Enhances inference speed on ARM64 devices by optimizing how CPU cores are used.
- **Resource Efficiency**: Ensures more efficient use of computational resources on devices with heterogeneous CPU cores, such as Raspberry Pi and Jetson platforms.