@lakshanthad 2024-06-09
@glenn-jocher it seems that the main branch does not have bad performance as @MarkhamLee mentioned and this PR does not differ results from `main` branch. Are you certain you got only 4FPS? 

I have tested on both RPi5 and RK3588. The below are on RK3588 for both `main` branch and this PR:

### Validation

```sh
(env) radxa@rock-5b:~/YOLOv8/ultralytics$ yolo val task=detect model=yolov8n_ncnn_model imgsz=640 data=coco8.yaml
Ultralytics YOLOv8.2.29 🚀 Python-3.9.2 torch-2.3.1 CPU (Cortex-A55)
Loading yolov8n_ncnn_model for NCNN inference...
Forcing batch=1 square inference (1,3,640,640) for non-PyTorch models
val: Scanning /home/radxa/YOLOv8/datasets/coco8/labels/val.cache... 4 images, 0 backgrounds, 0 corrupt: 100%|██████████| 4/4 [00:0
                 Class     Images  Instances      Box(P          R      mAP50  mAP50-95): 100%|██████████| 4/4 [00:00<00:00,  7.95
                   all          4         17      0.664      0.833      0.806      0.603
                person          3         10      0.722        0.5      0.527      0.268
                   dog          1          1       0.37          1      0.497      0.348
                 horse          1          2      0.777          1      0.995      0.747
              elephant          1          2      0.874        0.5      0.828      0.366
              umbrella          1          1      0.607          1      0.995      0.995
          potted plant          1          1      0.636          1      0.995      0.895
Speed: 7.4ms preprocess, 60.0ms inference, 0.0ms loss, 23.4ms postprocess per image
```

### Inference on webcam feed

```sh
(env) radxa@rock-5b:~/YOLOv8/ultralytics$ yolo detect predict model=yolov8n_ncnn_model source="1" show
Ultralytics YOLOv8.2.29 🚀 Python-3.9.2 torch-2.3.1 CPU (Cortex-A55)
Loading yolov8n_ncnn_model for NCNN inference...

1/1: 1... Success ✅ (inf frames of shape 640x480 at 30.00 FPS)

0: 640x640 1 person, 2 chairs, 2 beds, 94.4ms
0: 640x640 1 person, 2 chairs, 2 beds, 101.2ms
0: 640x640 1 person, 2 chairs, 2 beds, 73.0ms
0: 640x640 1 person, 2 chairs, 2 beds, 49.3ms
0: 640x640 1 person, 2 chairs, 2 beds, 59.3ms
0: 640x640 1 person, 2 chairs, 2 beds, 76.2ms
0: 640x640 1 person, 2 chairs, 2 beds, 50.6ms
0: 640x640 1 person, 2 chairs, 2 beds, 70.8ms
0: 640x640 1 person, 2 chairs, 2 beds, 60.2ms
0: 640x640 1 person, 2 chairs, 2 beds, 61.9ms
0: 640x640 1 person, 2 chairs, 2 beds, 60.4ms
0: 640x640 1 person, 2 chairs, 2 beds, 69.3ms
0: 640x640 1 person, 2 chairs, 2 beds, 64.3ms
0: 640x640 1 person, 2 chairs, 2 beds, 66.3ms
0: 640x640 1 person, 2 chairs, 2 beds, 62.5ms
0: 640x640 1 person, 2 chairs, 2 beds, 59.0ms
0: 640x640 1 person, 2 chairs, 2 beds, 77.5ms
0: 640x640 1 person, 2 chairs, 2 beds, 63.4ms
0: 640x640 1 person, 2 chairs, 2 beds, 67.1ms
0: 640x640 1 person, 2 chairs, 2 beds, 61.3ms
0: 640x640 1 person, 2 chairs, 2 beds, 67.4ms
0: 640x640 1 person, 2 chairs, 2 beds, 59.8ms
0: 640x640 1 person, 2 chairs, 2 beds, 64.0ms
0: 640x640 1 person, 2 chairs, 2 beds, 59.4ms
0: 640x640 1 person, 2 chairs, 2 beds, 59.9ms
0: 640x640 1 person, 2 chairs, 2 beds, 63.0ms
0: 640x640 1 person, 2 chairs, 2 beds, 64.7ms
0: 640x640 1 person, 2 chairs, 2 beds, 67.7ms
0: 640x640 1 person, 2 chairs, 2 beds, 64.7ms
0: 640x640 1 person, 2 chairs, 2 beds, 63.1ms
0: 640x640 1 person, 2 chairs, 2 beds, 64.5ms
0: 640x640 1 person, 2 chairs, 2 beds, 64.6ms
0: 640x640 1 person, 2 chairs, 2 beds, 62.7ms
0: 640x640 1 person, 2 chairs, 2 beds, 67.3ms
0: 640x640 1 person, 2 chairs, 2 beds, 61.6ms
0: 640x640 1 person, 2 chairs, 2 beds, 63.5ms
0: 640x640 1 person, 2 chairs, 2 beds, 63.4ms
```