@MarkhamLee 2024-06-10
After doing a lot more testing, I'd say this probably doesn't need to be a PR unless there are other use cases related to threads, instead some documentation notes:

* An environment that works for TF Lite will likely cause issues with NCNN and other formats/don't assume the same environment will support all export formats. 
* If things are running slow on a Rockchip 3588 device, check your core use in Htop, if you see all eight with heavy use, try rebuilding your environment as you should only see four in use when things are running well 

More details below....

@lakshanthad I have a several different environments, where I see the slow performance on my OrangePi 5Plus, running Ubuntu 22.04 across multiple OS installations (I have a couple of drives I swap in and out): 
`
(yolo_projects) markham@orangepi5plus-2204-node0:~/Documents/tests$ yolo val task=detect model=yolov8n_ncnn_model imgsz=640 data=coco8.yaml
Ultralytics YOLOv8.2.16 🚀 Python-3.10.14 torch-2.3.0 CPU (Cortex-A55)
Loading yolov8n_ncnn_model for NCNN inference...
Forcing batch=1 square inference (1,3,640,640) for non-PyTorch models
val: Scanning /home/markham/Documents/GitHub/computer-vision-demo-reel/orangepi5plus_ncnn_detect/datasets/coco8/labels/val.cache... 4 images, 0
                 Class     Images  Instances      Box(P          R      mAP50  mAP50-95): 100%|██████████| 4/4 [00:02<00:00,  1.78it/s]
                   all          4         17      0.664      0.833      0.806      0.603
                person          4         10      0.722        0.5      0.527      0.268
                   dog          4          1       0.37          1      0.497      0.348
                 horse          4          2      0.777          1      0.995      0.747
              elephant          4          2      0.874        0.5      0.828      0.366
              umbrella          4          1      0.607          1      0.995      0.995
          potted plant          4          1      0.636          1      0.995      0.895
Speed: 9.0ms preprocess, 251.6ms inference, 0.0ms loss, 252.4ms postprocess per image
Results saved to runs/detect/val3
💡 Learn more at https://docs.ultralytics.com/modes/val
(yolo_projects) markham@orangepi5plus-2204-node0:~/Documents/tests$ 
`

After doing more testing, I "think" this is actually a dependency issue that arises when you adjust dependencies so that TF Lite exports can run successfully. I.e., I always adjust my environments to be able to run TF Lite, but when I created fresh environments and didn't bother with TF after seeing your results, I got similar results as you. Also worth noting that my benchmark results in an environment where TF Lite fails to run are faster across the board:

With TF Lite able to run 
![Screenshot from 2024-06-09 20-12-59](https://github.com/ultralytics/ultralytics/assets/16950894/07300956-3ae0-4459-886e-1d674ae7a17a)


Without TF Lite 
![Screenshot from 2024-06-09 16-53-36](https://github.com/ultralytics/ultralytics/assets/16950894/389a81a4-f4f4-4299-b533-ff64aa4a154e)

I saw the faster results with versions 8.2.15 and 8.2.29, so I'm fairly sure it's a TF dependencies issue and not related to Ultralytics.

On the thread/core front, I did notice that when things ran slow, all eight cores saw heavy use:

![Screenshot from 2024-06-09 18-57-30](https://github.com/ultralytics/ultralytics/assets/16950894/5382d68a-f4ec-4e76-a3b9-efa42865b059)

When things run fast, only four cores see heavy use:
![Screenshot from 2024-06-09 19-00-55](https://github.com/ultralytics/ultralytics/assets/16950894/97554a52-9e23-45e4-9eb0-9695a9230e4d)

I did deploy the current GH repo to a fresh environment in editable mode and experiment with changing the number of threads, not passing that parameter or specifying four threads was usually about the same, specifying one thread, two, six, eight, etc., was always slower. Seems my intuition was right about the relevance of threads, just wrong about the the root cause. It also means that "something" under the hood is specifically targeting the four "big" cores.